+++ sample/ChatSample/Client/Client.csproj
