+++ sample/ChatSample/Client/Program.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Hosting;
 
-namespace ChatSample
+namespace Client
 {
     public class Program
     {
