+++ sample/ChatSample/Client/wwwroot/lib/signalr.js
